# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for quiz game with host management and Telegram bot integration",
        "title": "Quiz Game API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/sessions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start a new quiz session, generates a join code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Create a quiz session",
                "parameters": [
                    {
                        "description": "Session data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.CreateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/quiz-game-backend_internal_services.SessionState"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sessions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current state of a quiz session including current question",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Get session state",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz-game-backend_internal_services.SessionState"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sessions/{id}/leaderboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get session leaderboard sorted by score",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Get leaderboard",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/quiz-game-backend_internal_services.LeaderboardEntry"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sessions/{id}/next": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start quiz or advance to the next question. Finishes quiz if no more questions.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Move to next question",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz-game-backend_internal_services.SessionState"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{id}/reveal": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Show correct answer and calculate scores for current question",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Reveal the correct answer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz-game-backend_internal_services.SessionState"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ws/session/{id}": {
            "get": {
                "description": "Connect via WebSocket to receive real-time session updates",
                "tags": [
                    "websocket"
                ],
                "summary": "WebSocket connection for session updates",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "internal_handlers.CreateSessionRequest": {
            "type": "object",
            "required": [
                "quiz_id"
            ],
            "properties": {
                "quiz_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
...
        "quiz-game-backend_internal_models.Option": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "question_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "quiz-game-backend_internal_models.Participant": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "joined_at": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "session_id": {
                    "type": "integer"
                },
                "telegram_id": {
                    "type": "integer"
                },
                "total_score": {
                    "type": "integer"
                }
            }
        },
        "quiz-game-backend_internal_models.Question": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quiz-game-backend_internal_models.Option"
                    }
                },
                "order_num": {
                    "type": "integer"
                },
                "quiz_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "quiz-game-backend_internal_models.Quiz": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "host_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quiz-game-backend_internal_models.Question"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "quiz-game-backend_internal_services.LeaderboardEntry": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "telegram_id": {
                    "type": "integer"
                },
                "total_score": {
                    "type": "integer"
                }
            }
        },
...
        "quiz-game-backend_internal_services.OptionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_correct": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "quiz-game-backend_internal_services.QuestionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quiz-game-backend_internal_services.OptionResponse"
                    }
                },
                "order_num": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "quiz-game-backend_internal_services.SessionState": {
            "type": "object",
            "properties": {
                "answer_count": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_question": {
                    "type": "integer"
                },
                "current_question_data": {
                    "$ref": "#/definitions/quiz-game-backend_internal_services.QuestionResponse"
                },
                "host_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quiz-game-backend_internal_models.Participant"
                    }
                },
                "quiz": {
                    "$ref": "#/definitions/quiz-game-backend_internal_models.Quiz"
                },
                "quiz_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_questions": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}